requirements.txt
Appium-Python-Client==4.4.0
attrs==24.3.0
browserstack-local==1.2.10
browserstack_sdk==1.23.5
certifi==2024.12.14
charset-normalizer==3.4.1
dukpy==0.5.0
exceptiongroup==1.2.2
execnet==2.1.1
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
percy-appium-app==2.0.6
pluggy==1.5.0
psutil==6.1.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-SystemConfiguration==11.0
PyPAC==0.16.5
PySocks==1.7.1
pytest==8.3.4
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.27.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
